think these will work here if setting/getting state variables
