+++ src/constitutive/mod.rs
@@ -3,16 +3,43 @@ mod test;
 
 mod hyperelastic;
 
-use crate::mechanics::
+use crate::
 {
-    DeformationGradient,
-    LeftCauchyGreenDeformation,
-    Scalar
+    math::
+    {
+        ContractSecondIndexWithFirstIndexOf,
+        TensorRank2Trait,
+        TensorRank4Trait
+    },
+    mechanics::
+    {
+        CauchyStress,
+        CauchyTangentStiffness,
+        DeformationGradient,
+        FirstPiolaKirchoffStress,
+        FirstPiolaKirchoffTangentStiffness,
+        LeftCauchyGreenDeformation,
+        Scalar
+    }
 };
 
 pub type ConstitutiveModelParameters<'a> = &'a [Scalar];
 
-pub trait ConstitutiveModel
+pub trait ConstitutiveModel<'a>
 {
+    fn calculate_cauchy_stress(&self, deformation_gradient: &DeformationGradient) -> CauchyStress;
+    fn calculate_cauchy_tangent_stiffness(&self, deformation_gradient: &DeformationGradient) -> CauchyTangentStiffness;
     fn calculate_left_cauchy_green_deformation(&self, deformation_gradient: &DeformationGradient) -> LeftCauchyGreenDeformation;
+    fn calculate_first_piola_kirchoff_stress(&self, deformation_gradient: &DeformationGradient) -> FirstPiolaKirchoffStress
+    {
+        self.calculate_cauchy_stress(deformation_gradient)*deformation_gradient.inverse_transpose()*deformation_gradient.determinant()
+    }
+    fn calculate_first_piola_kirchoff_tangent_stiffness(&self, deformation_gradient: &DeformationGradient) -> FirstPiolaKirchoffTangentStiffness
+    {
+        let deformation_gradient_inverse_transpose = deformation_gradient.inverse_transpose();
+        let first_piola_kirchoff_stress = self.calculate_first_piola_kirchoff_stress(deformation_gradient);
+        self.calculate_cauchy_tangent_stiffness(deformation_gradient).contract_second_index_with_first_index_of(&deformation_gradient_inverse_transpose)*deformation_gradient.determinant() + FirstPiolaKirchoffTangentStiffness::dyad_ij_kl(&first_piola_kirchoff_stress, &deformation_gradient_inverse_transpose) - FirstPiolaKirchoffTangentStiffness::dyad_il_kj(&first_piola_kirchoff_stress, &deformation_gradient_inverse_transpose)
+    }
+    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar;
+    fn new(parameters: ConstitutiveModelParameters<'a>) -> Self;
 }
